energy_forecast/requirements.txt
# Core Data Processing
numpy>=1.21.0  # Array processing
pandas>=1.3.0  # Data manipulation
scipy>=1.7.0   # Scientific computing

# Machine Learning
scikit-learn>=1.0.0  # Traditional ML algorithms
tensorflow>=2.8.0    # Deep learning
lightgbm>=3.3.0     # Gradient boosting
xgboost>=1.5.0      # Extreme gradient boosting
optuna>=2.10.0      # Hyperparameter optimization

# Visualization
plotly>=5.3.0       # Interactive plots
matplotlib>=3.4.0   # Static plots
seaborn>=0.11.0     # Statistical visualizations

# Time Series Analysis
statsmodels>=0.13.0  # Statistical models
prophet>=1.0.0       # Time series forecasting

# Utilities
tqdm>=4.62.0        # Progress bars
joblib>=1.1.0       # Parallel processing
psutil>=5.8.0       # System monitoring

# Testing
pytest>=6.2.5       # Testing framework
pytest-cov>=2.12.0  # Coverage reporting

# Development
black>=21.7b0       # Code formatting
isort>=5.9.0        # Import sorting
flake8>=3.9.0       # Code linting
mypy>=0.910         # Type checking

# Documentation
sphinx>=4.1.0       # Documentation generator
sphinx-rtd-theme>=0.5.2  # Documentation theme

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart==0.0.6
aiofiles>=0.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL support

# Logging and Monitoring
python-json-logger>=2.0.0  # JSON logging
prometheus-client>=0.11.0  # Metrics collection
